it appears that the memory leak from tf 2.4.0 which doesnt happen in tf2.2.0, is still happening in tf 2.3.0. or maybe its a feature?
