+++ utilities.py
@@ -66,87 +66,3 @@ def log(text):
     if verbose:
         print(text)
 
-
-
-def do_inference(model, epoch):
-
-    #todo: this is aweful
-
-    if epoch is None:
-        epoch = 'init'
-    else:
-        epoch = epoch + 1
-
-    datagenerator = model.datagenerator
-
-    if model.loaded == False:
-        #todo: test if this condition ever actually happens
-        log('do_inference recieved unloaded model')
-        model = load_model()
-    else:
-        log('got loaded model for inference')
-
-    log(type(model))
-
-    ## just in case this hasnt already been done
-    model.build(input_shape=(None, 64, 64, 3))
-
-    #idk why
-    if epoch is None:
-        model.summary()
-
-    #get a batch of images from the datagenerator
-    images, labels = datagenerator.next()
-
-    ## make an array of input images
-    in_images =[]
-    i = 0
-    for item in images:
-
-        pil_img = tf.keras.preprocessing.image.array_to_img(item)
-        log(type(pil_img))
-
-        dir = 'ae_samples/'
-
-        if not os.path.exists(dir):
-            os.makedirs(dir)
-
-        in_images.append(pil_img)
-        i = i + 1
-
-    ## do the inference on the inputs
-    result = model.predict(images, batch_size=batch_size)
-
-    ## make an array of output images
-    out_images = []
-    i = 0
-    for item in result:
-        img = item
-        img = img[:, :, 0:3] #/ 127.0
-
-        pil_img = tf.keras.preprocessing.image.array_to_img(img)
-        out_images.append(pil_img)
-        i = i + 1
-
-    #concatenate the two arays of images and add to list
-    a = 1
-    concatted = []
-    for item in zip(in_images, out_images):
-        #get_concat_h(item[0], item[1]).save(dir + '/ae_sample-' + str(a) + '.jpeg')
-        new_img = get_concat_h(item[0], item[1])
-        concatted.append(new_img)
-        a = a + 1
-
-    #concatenate images into rows 4 sets wide
-    rows = []
-    for i in range(0, len(concatted)-1, 4):
-        rows.append(concat_four_h(concatted[i], concatted[i+1], concatted[i+2], concatted[i+3]))
-
-    #concatenate the rows vertically t oform the final image
-    first_img = rows[0]
-    for i in range(1, len(rows)):
-        first_img = get_concat_v(first_img, rows[i])
-
-    #save the image to disk
-    first_img.save('ae_samples/epoch_' + str(epoch) + '.jpeg')
-    #return first_img
